+++ Source/Programs/Files.h
@@ -30,7 +30,7 @@ class FileExplorer {
 		std::vector<ClickBox*> folderElements;
 		std::vector<ClickBox*> folderElementsNames;
 
-		Button* selectedBox = nullptr; // Background hoover to selected element
+		Label* selectedBox = nullptr; // Background hoover to selected element
 
 		ClickBox* selectedElement = nullptr; // Reference to already selected folder element